Python/TG_Bots/AI_Bot-2.0/requirements.txt
aiogram>=3.22.0
python-dotenv>=1.0.0
g4f>=6.5.7
